change every smart contract has its own module
