+++ .github/workflows/linter.yml
@@ -24,7 +24,9 @@ jobs:
           - examples/cis2-multi/Cargo.toml
           - examples/cis2-nft/Cargo.toml
           - examples/cis2-wccd/Cargo.toml
-          - examples/cis2-wccd-upgradable/Cargo.toml
+          - examples/cis2-wccd-upgradable/proxy/Cargo.toml
+          - examples/cis2-wccd-upgradable/implementation/Cargo.toml
+          - examples/cis2-wccd-upgradable/state/Cargo.toml
           - examples/fib/Cargo.toml
           - examples/icecream/Cargo.toml
           - examples/memo/Cargo.toml
@@ -254,7 +256,9 @@ jobs:
         crates:
           - examples/two-step-transfer/Cargo.toml
           - examples/cis2-wccd/Cargo.toml
-          - examples/cis2-wccd-upgradable/Cargo.toml
+          - examples/cis2-wccd-upgradable/proxy/Cargo.toml
+          - examples/cis2-wccd-upgradable/implementation/Cargo.toml
+          - examples/cis2-wccd-upgradable/state/Cargo.toml
           - examples/cis2-nft/Cargo.toml
           - examples/cis2-multi/Cargo.toml
 
@@ -332,7 +336,9 @@ jobs:
           - examples/cis2-multi/Cargo.toml
           - examples/cis2-nft/Cargo.toml
           - examples/cis2-wccd/Cargo.toml
-          - examples/cis2-wccd-upgradable/Cargo.toml
+          - examples/cis2-wccd-upgradable/proxy/Cargo.toml
+          - examples/cis2-wccd-upgradable/implementation/Cargo.toml
+          - examples/cis2-wccd-upgradable/state/Cargo.toml
           - examples/fib/Cargo.toml
           - examples/icecream/Cargo.toml
           - examples/memo/Cargo.toml
@@ -453,7 +459,9 @@ jobs:
           - examples/cis2-multi/Cargo.toml
           - examples/cis2-nft/Cargo.toml
           - examples/cis2-wccd/Cargo.toml
-          - examples/cis2-wccd-upgradable/Cargo.toml
+          - examples/cis2-wccd-upgradable/proxy/Cargo.toml
+          - examples/cis2-wccd-upgradable/implementation/Cargo.toml
+          - examples/cis2-wccd-upgradable/state/Cargo.toml
           - examples/fib/Cargo.toml
           - examples/icecream/Cargo.toml
           - examples/memo/Cargo.toml
