+++ examples/cis2-wccd-upgradable/implementation/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "cis2_wccd_upgradable_implementation"
+version = "0.1.0"
+authors = ["Concordium <developers@concordium.com>"]
+edition = "2018"
+license = "MPL-2.0"
+
+[features]
+default = ["std"]
+std = ["concordium-std/std", "concordium-cis2/std"]
+
+[dependencies]
+concordium-std = {path = "../../../concordium-std", default-features = false}
+concordium-cis2 = {path = "../../../concordium-cis2", default-features = false}
+
+[lib]
+crate-type=["cdylib", "rlib"]
+
+[profile.release]
+opt-level = "s"
+codegen-units = 1
